Rename cpo/dom.py to cpo/dom_mgr.py
